Implementação Fluent Validations

+++ Granito.Calculo.Api/Application/Calculos/Commands/CalcularJurosCommand.cs
@@ -2,4 +2,4 @@
 
 namespace Granito.Calculo.Api.Application.Calculos.Commands;
 
-public sealed record CalcularJurosCompostoCommand(decimal ValorInicial, decimal Meses) : ICommand<decimal>;
+public sealed record CalcularJurosCompostoCommand(decimal ValorInicial, int Meses) : ICommand<decimal>;
